+++ DataBase/Car.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace DataBaseLib
 {
-    public class Car
+    public class Car : IQuery
     {
         public int Id { get; set; }
         public string RegSign { get; set; }
@@ -15,5 +16,14 @@ namespace DataBaseLib
         public DateTime Year { get; set; }
         public int Mileage { get; set; }
         public string Colour { get; set; }
+
+        public void Insert(ConnectionDB connection)
+        {
+        }
+
+        public IList Select(ConnectionDB connection)
+        {
+            return new List<Car>();
+        }
     }
 }
