+++ DataBase/Service.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -6,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace DataBaseLib
 {
-    class Service
+    public class Service : IQuery
     {
         public int Id { get; set; }
         public int IdCar { get; set; }
@@ -14,5 +15,15 @@ namespace DataBaseLib
         public bool ControlBreak { get; set; }
         public int Antifreeze { get; set; }
         public bool Tires { get; set; }
+
+        public void Insert(ConnectionDB connection)
+        {
+         
+        }
+
+        public IList Select(ConnectionDB connection)
+        {
+            return new List<Car>();
+        }
     }
 }
